♻️ [RUMF-1209] frustration detection groundwork
This commit changes the design of `trackActions` to prepare for
frustration detection. In particular, it re-introduce the concept of
"potential action" that will be used when generating rage clicks.

It also introduces a concept of `TrackActionState` to be able to split
the implementation into multiple functions, as the `trackActions`
implementation will grow a bit.