⚗✨ [RUMF-1209] introduce "dead" and "error" frustration types (#1487)
* ♻️✅ [RUMF-1209] merge trackActions tests

It makes no sense anymore to split trackActions into two `describe`,
since the `newAction` function does not exist anymore, and tests are
mostly similar. Let's merge the two `describe` and reorder a few test.

* ♻️ [RUMF-1209] frustration detection groundwork

This commit changes the design of `trackActions` to prepare for
frustration detection. In particular, it re-introduce the concept of
"potential action" that will be used when generating rage clicks.

It also introduces a concept of `TrackActionState` to be able to split
the implementation into multiple functions, as the `trackActions`
implementation will grow a bit.

* ✨ [RUMF-1209] introduce "dead" and "error" frustration types

* ✨ [RUMF-1209] add frustration_types on actions events

* ✅ [RUMF-1209] add frustration related tests to trackActions

* 👌 replace 'a action' to 'an action'

* 👌 remove TrackActionsState, replace with closures

* 👌🐛 do not set a duration for dead clicks

* 👌 rename endClick to stopClickProcessing

* 👌 simplify PotentialAction

* Merge `notifyIfComplete` and `complete` into a new `validate` method
* Remove unneeded `finalState`
* Factorize cleanup code to a `stop` function
* Remove unused `base` property
* Implement a `addFrustration` method

* ✅👌 add e2e tests

* specialize `trackAction` to `trackClickActions`

* unify action namings

action -> click action
potential action -> potential click action

* 👌✅ add a e2e test for multiple frustrations

* 👌 more explicit type for id history

* 👌 replace unnecessary call to isExperimentalFeatureEnabled

* 👌 use a helper to transform a Set to an Array

* 🏷️ [RUMF-1209] add "ES2015" to the test app tsconfig

We now use Set to track actions, and it appears in module type
definitions (.d.ts). By default, TS does not include the type definition
for Set, so we need to use the `ES2015` lib.

This should not be a breaking change, since we already used Set in other
places of the SDK.