♻️ [RUMF-1267] remove circular dependencies part 1 (#1559)
* 🚨 [RUMF-1267] enable import/no-cycle as a warning

* ♻️ [RUMF-1267] split some imports

By splitting imports, we import less stuff, which removes a few cycles

* ♻️ [RUMF-1267] remove cycle in recorder

Move `shouldIgnoreNode` to `privacy` module. `shouldIgnoreNode` is only
used in `privacy`, so it makes sense to put it there.

* ♻️ [RUMF-1267] remove cycle in session

Move session constants to a separate module

* ♻️ [RUMF-1267] remove cycle in performances script

Move the utility function to check if an URL is a SDK bundle URL into a
new utility module.

* ♻️ [RUMF-1267] remove cycle in e2e tests

Move `flushEvents` in the `framework`, as it is both using `framework`
functions and is used by the `framework`.