🐛 improve `jsonStringify` implementation (#1653)
* 🐛 improve `jsonStringify` implementation

The previous implementation only removed toJSON on the input value and
its prototype. We usually pass an Object as input, so it was focused
on Object.

But websites using outdated libraries that define toJSON to Array
objects (ex: prototype.js before 1.7) are still an issue with this
strategy.

This new implementation assumes that internally, the SDK won't use other
objects than Array and Object (primitive types like string and number
are unaffected), and those objects won't be defined with an own toJSON
method. So, it focuses on removing toJSON to Array and Object
prototypes, which completely solves the issue.

* 👌🐛 keep removing toJSON from the root value

* 👌 rename 'remove' to 'detach'

* 🐛 fix IE11 support

* 👌 apply review suggestions