✨ [RUMF-1396] add a Disconnect alert
With Manifest V3, connection between the devtools panel and the
background service worker can be lost when the service worker is
refreshed. I experienced it very often while working on the extension
(every time I changed the background script). To help UX/DX, this commit
adds an alert to allow refreshing easily the devtools panel.