⬆️ [RUMF-1434] fix yarn failing to install puppeteer on M1 mac (#1843)
* ⬆️ [RUMF-1434] upgrade puppeteer

* 🐛 [RUMF-1434] workaround network tracking losing track of requests

Sometimes, puppeteer (or the underlying CDP session) don't notify us
when a request finishes. This makes the scenario stuck indefinitely.

This commit improves the situation by:

* listening to the `response` event (it has been observed that
  sometimes `requestfailed` and `requestfinished` are not notified, but
  `response` is)

* adding a 5 seconds timeout to all requests

* 🐛 [RUMF-1434] fix CPU profiling

It seems like following the puppeteer upgrade (maybe because it uses a
different Chromium version), the Chrome Devtools `Profiler` profile is
reset on each page navigation. Thus, we lose some precious information.

This commit fix it by stop/start the profiler more often, aggregating
the different resulting profiles.

* 🐛 [RUMF-1434] fix twitter scenario

Twitter does not display the "Preferences" page controls if we don't
consent to cookies anymore. This commit adds the expected cookie to
opt-in to cookie consent.

* 👌 add doc