♻️ use different error handling strategy for background / devtools panel
Communication strategies are different between the background script and
the devtools panel.

For the background script, we don't mind if we don't have a receiver
when sending messages (for example, the devtools panel has been closed,
so it's not a big deal).

For the devtools panel, we should be extra cautious when communication
fails because the background script disconnected. We'll use the
now generic `disconnect` module in other places in the future.