✨ Move extension settings to a Settings panel (#1847)
* ⬆️ upgrade mantine

To use the new features of Mantine, let's upgrade to the latest version.

* 💄 remove margin on body

This allows the inner components to chose where to put margins. It also
helps to implement a nice scroll strategy for tabs because now we can
use things like `100vh`.

* ♻️ introduce a Columns component

This component is extracted from the Infos tab and will be reused in the
future Settings tab.

* ✨ implement a Settings tab

Settings are getting a bit dense. Let's use a separate Settings tab for
it. This is a bit similar to what Chrome and Firefox devtools are doing,
in the sense that settings are centralized in a full screen interface.

* 💄 improve integration

This commit improves the overall layout and make tabs content scroll as
expected.