👷 change TS target to esnext for the extension
Targeting es5 isn't needed in chrome extensions. Using a newer target
makes it simpler to debug async code. I chose 'esnext' so TS will
transpile the least amount of things. Since we don't use anything fancy,
it should be fine for the time being. We can always reduce this value
if/when we use a syntax that is not yet supported in Chrome.