✨ send data from the SDK to the extension (#1850)
* ✨ send any SDK data to the extension

* 🔊 create a logger module

Logs within the extension are getting hard to follow. Let's use a logger
module to keep things manageable.

* ♻️ use different error handling strategy for background / devtools panel

Communication strategies are different between the background script and
the devtools panel.

For the background script, we don't mind if we don't have a receiver
when sending messages (for example, the devtools panel has been closed,
so it's not a big deal).

For the devtools panel, we should be extra cautious when communication
fails because the background script disconnected. We'll use the
now generic `disconnect` module in other places in the future.

* ✨ forward messages from the SDK up to the devtools panel

That was harder that expected.

* ✨ display events from the SDK

* ✅ use a different SDK version during unit tests

IE11 does not support custom events. Instead of adding runtime code to
check if the browser support them, just use a different SDK version for
tests so we can distinguish between dev builds and test runs.

* 🐛 fix dev bundle redirection for -canary and -staging suffix

* ♻️ simplify a bit the background connection protocol

No need to have an `init` message anymore, each connection send the
tabId inside the port name.

* 👷 change TS target to esnext for the extension

Targeting es5 isn't needed in chrome extensions. Using a newer target
makes it simpler to debug async code. I chose 'esnext' so TS will
transpile the least amount of things. Since we don't use anything fancy,
it should be fine for the time being. We can always reduce this value
if/when we use a syntax that is not yet supported in Chrome.

* 🚸 always forward records to the extension

The goal of this commit is to allow sending data to the extension
without being in a dev build. The constraints are:

* have a minimal impact on SDK performances and bundle size when the
  extension isn't used

* collect all events reliably, even those that are sent during SDK
  `init` in a synchronous setup.

To achieve this, this commit:

* changes the way the SDK is sending messages to the extension from a
  custom event (unknown side effects, hard to evaluate performance
  impact) to a simple call to an optional function callback on `window`.

* because only content scripts are guaranteed to be executed before any
  script on the page, it uses the new `scripting` webextension API to
  register two (yes, two!) content scripts:

  * one executing in the "main" execution world (= the webpage),
    defining the callback function on `window`. This function will
    simply dispatch the custom event that used to be dispatched by the
    SDK before

  * one executing in the "isolated" execution world, listening to those
    custom events and sending them to the background script.

In order to have a minimal impact on normal RUM/Logs processing, those
content scripts are only active while the devtools panel is open.

* 📝 change Settings tab doc

In the future, I'd like to make this the default event source if the SDK
matches a given version. For that, I first need to release it so we know
against which version to test.