👷 fix publish-npm job (#1859)
The publish-npm job is only using the cache and does not install yarn
dependencies itself. Because the cache is shared among all other job, it
can be updated with other dependency versions in the meantime, and the
publish-npm job fails.

This commit fixes the issue by running `yarn`, ensuring the right
dependencies are installed before running the publish-npm script.