🐛 move document_version to global context
It turns out that the View `_dd.document_version` is replaced by another
unrelated version in the backend. I wanted to duplicate this field so we
can keep the original in views.

Because unknown `_dd` fields are discarded from View events by the
backend, I chose to put it in the global context, similarly to #1757
(and other before).