⚗🐛 [RUMF-1296] use pointerup to trigger click actions  (#1958)
* ♻️ [RUMF-1296] simplify a bit how we compute input user activity

This commit will help implementing click actions triggered by pointerup
behind a FF.

* ♻️✅ simplify trackClickActions tests

* 🐛 [RUMF-1296] ignore non-primary pointer events

In 'listenActionEvents', we expect that a 'click' event is preceded by
a single 'pointerdown' event. It turns out this is not always the case
on multitouch devices.

This issue would be amplified when listening to 'pointerup' events.

Let's just focus on the 'primary' pointer events. This will ignore
secondary touches until we implement proper support for them (if ever).

* ⚗🐛 [RUMF-1296] use pointerup to trigger click actions

This slightly change the action beginning, but takes any activity
produced by 'pointerup' 'mouseup' 'touchend' events into account. If
those actions don't produce any activity, the action start is likely to
be the same as before, since 'pointerup/mouseup/touchend' should be
triggered very close to the 'click' event.

* 🔊⚗ [RUMF-1296] record the delay between pointerup and click

* 🚩 rename feature flag

* 🐛 restore existing behavior for dead clicks

* 👌 rename onActionStart to onStartEvent