🔇 remove some errors during Worker creation from telemetry (#2035)
When failing to create the Worker, we try to detect whether the error is
related to CSP to guide the user.

When using the `require-trusted-types-for` CSP, the error does not
mention "Content Security Policy". Thus, we report it as a telemetry
error.

This commit adjusts this and consider it as a CSP error.