👷 extend Volta configuration in most package.json (#2048)
* 👷 extend Volta configuration in most package.json

With this change, the right version of Node and Yarn will be used when
invoking commands from sub-packages directories. This is particularly
useful with "native"/binary dependencies installed for a given Node
version which will not run with a different version.

Ex: running `yarn dev` in developer-extension/, or `yarn start` in
performances/.

* 👷 remove nodenv support

Everyone should be use Volta by now.