♻️ [RUMF-1534] collocate view updates
This commit moves all view updates within the "newView" function.
Motivations:

* This helps to track down when a view can be updated.
* It makes sure that we don't send a view update on page unload when the
  session did expire before.

One important functional change: a view update is now always sent when
ending a view

It was not the case before: we intentionally didn't send a view update
when the session was renewed, because it could lead to imprecise
measurements (ex: inaccurate view duration), especially when the page
was frozen for a long time. But even if we didn't send a view update in
this case, there was a good chance that a view update was still sent
because of other events (ex: periodic session keepalive, counter
updates...)

We now always send a view update when the session is expired, which is
a bit less problematic than when the session is renewed (because it
happens sooner, without relying on arbitrary user interaction), but
might still suffer from imprecise measurements in case of page freeze.
We think this is an acceptable drawback and should not have a
perceptible impact.

This is an interesting change because next we'll want to send a view
update to let the backend know that the session has expired.