🐛 [RUMF-1436] patch History.prototype.* instead of history.*
This is not strictly needed, but to keep things simple, the previous
patch only cared on the *own property* descriptor. So if the method that
we want to instrument is not directly on the object, the test is
ineffective.

For history methods, we were using the `window.history` instance, but
methods are actually defined on its prototype, so we wouldn't be able to
detect whether those methods are readonly.