🐛 always serialize <style> element state
Previously, we serialized the <style> element state (as reflected by its
associated StyleSheet object) only if the element was empty / did not
have children.

This is an issue when the recording starts after a non-empty <style>
element is edited via `insertRule`/`deleteRule`, as doing so only
changes the <style> element state and not its children, so changes were
ignored.

This commit fixes this issue by always serializing the <style> element
state.