♻️ store requests instead of events
To allow tests related to requests instead of events, let's store
requests directly in the IntakeRegistry.

The `IntakeRequest` model is reflects what's sent to the intake, for
example we send telemetry events to the rum intake. This avoids awkward
concepts like a "telemetry intake".

Also, collocate type assertion functions since it's only used in this
module.