🐛 fix dynamic style serialization (#2397)
* ✅ revamp serialization tests a bit

Some tests were incorrectly labeled and hard to grasp. This commit
improves a bit, and adds a test on how style element containing text are
currently serialized.

* 🐛 always serialize <style> element state

Previously, we serialized the <style> element state (as reflected by its
associated StyleSheet object) only if the element was empty / did not
have children.

This is an issue when the recording starts after a non-empty <style>
element is edited via `insertRule`/`deleteRule`, as doing so only
changes the <style> element state and not its children, so changes were
ignored.

This commit fixes this issue by always serializing the <style> element
state.

* ⚡️ do not serialize <style> children

Because the <style> state is now always serialized in the _cssText
attribute, providing children is not useful and implies sending the CSS
text twice in some cases.

* ⚡️remove isStyle checks and property

When serializing text nodes, there is no need to check whether the
parent element is a <style> since we don't serialize <style> element
children anymore.

The `isStyle` property on the `TextNode` interface is kept for backward
compatibility but won't be defined for newer versions of the SDK.

* ✅ fix tests on Safari 14.1 and Firefox

Browsers are serializing some CSS rules slightly differently. Let's use
CSS that is serialized the same every where.

* 👌 remove outdated comment