⚗️✨ [RUM-253] compress RUM data (#2400)
* ✨ [RUM-253] introduce an abstract Encoder interface

This new interface is slightly different than the current deflate
encoder implementation. This is because it has been adjusted to take the
Batch constraints into account.

The deflate encoder implementation will be changed in a future commit.

* ✨ [RUM-253] use an Encoder in Batch

* ♻️ [RUM-253] adjust DeflateEncoder implementation

* 🚚 [RUM-253] move some deflate types to core

* [RUM-253] allow providing an encoder when creating batches

* ⚗✨ [RUM-253] use deflate encoders for RUM batches

This commit is not pretty. To avoid having inflated packages while
dogfooding, we pass some functions from rum to rum-core when creating
the RUM public API.

In the future, those functions will be moved to the core package, and
rum-core will be able to import them directly.

* 🐛 fix support for IE

* 👌 clearer variable name

* 👌 cancel finish callbacks when running finishSync()

* 👌 use the Worker instance created from RUM in the recorder

* 👌 simplify encoding creation condition

* [RUM-253] increase a bit the estimated deflate ratio + add justification