📦 update mantine usages
See https://mantine.dev/guides/6x-to-7x/

Main changes:

* `sx` props and a few JS APIs related to styling have been replaced by
  a css module approach.

* a few props have been renamed (ex: `color=` to `c=`)

* the Table component is now exposing related components (ex:
  `Table.Td`) instead of using intrinsic elements (ex: `td`) directly.

Note: contrary to mantine [recommended setup][1], I did not added
postcss as a dependency. Using postcss could bring some [extra
capabilities][2] that are not yet implemented in Chromium but to me they
are not worth the extra dependency/setup for now.

[1]: https://mantine.dev/getting-started/#get-started-without-framework
[2]: https://mantine.dev/styles/postcss-preset/