♻️ pass the Observable instance to the onFirstSubscribe callback
This change mimics the [TC39 observable proposal][1] (and various
implementations) by providing a way to notify directly as a
onFirstSubscribe parameter.

Contrary to the TC39 proposal and other implementations, it simply
provide the observable instance, instead of introducing a new "observer"
concept (Observable and Observer concepts are "merged" in our
implementation).

This is a simple quality of life improvement, but also fixes a typing
issue I stumbled upon where TS claimed that the "observable" variable
was not necessary defined when it referred to the outer scope variable.

[1]: https://github.com/tc39/proposal-observable