🔉 [RUM-253] adjust heavy customer data warning when compression is enabled (#2529)
* ♻️ [RUM-253] refactor customer data tracking

Use a single component used by both feature flags and context manager.

* ♻️ [RUM-253] refactor tests

No need to test for customer data warning in every spec, move related
tests to `trackCustomerData` spec.

* ♻️ [RUM-253] expose customerDataTracker in ContextManager

This allows to reuse the same customerDataTracker in the "stored context
manager" logic, so we don't show the same warning twice. It simplifies a
bit the code.

* 🔊 [RUM-253] adjust warning limit when compression is enabled

Have a bigger limit when the compression is enabled. Because the
compression "status" is only known at init, we need to wait until before
showing the warning.

* 👌 rename module

* 👌 rename `maybeWarn` to something more explicit

* 👌 adjust documentation URL

* 👌 use a "customer data tracker manager"

* 👌 use global limit for all contexts

* ♻️ do not store data type in trackers

* ♻️ simplify customer data telemetry