♻️ [RUM-2445] use recorderApi directly (#2563)
The goal of this commit is to simplify the rumPublicApi module by using
the `recorderApi` directly, without buffering the calls.

This is possible because the recorder api already has logic to support
early calls to `start` and `stop`. The `getSessionReplayLink`
implementation had to be slightly changed to work similarly to `start`
and `stop`.