🐛 [RUM-2782] validate resource timings more granularly
Before this commit, we dropped any `PerformanceResourceEntry` that had
inconsistent timings. With this commit, when the experimental feature
"tolerant_resource_timings" is enabled, the various timings are
validated independently.

This should improve the situation on Chrome when a Service Worker is
involved. In this case, most of the time, connection timings are
unexpectedly set after the `requestStart` timing (see Chromium
issue[1]). Before this change, the whole `PerformanceResourceEntry`
would be ignored, and no timings would be included. With this change
(and the flag enabled), relevant timings will be correctly defined.

Related issue: #2566

[1]: https://issues.chromium.org/issues/323703325