update aws-cli setup
python2 (aka 'python' on Debian) is not available anymore in Bookworm.
So we would need to upgrade to python3, but the only thing that uses
python is aws-cli, and newer version of aws-cli includes its own python
version. So let's remove python altogether and use the newer version of
aws-cli instead.