🐛 [RUM-5785] fix track navigation timings on Safari
In most browsers, when observing `PerformanceNavigationTiming` via a
`PerformanceObserver` before the page is fully loaded, the same
"navigation" entry is notified twice:

* once with the "buffered" entry, even if it is incomplete

* once when all the timings are defined, after the `load` event is
  complete.

On Safari, the entry is only notified once. The SDK is only aware of the
incomplete entry (which is discarded), so all timings related to the
navigation entry are missing.

To work around this issue and keep things simple, this commit removes
the `PerformanceObserver` usage. Instead, we wait for the `load` event
to be complete, and then just pick up the buffered navigation entry.
This is similar to the strategy we use to generate the "document"
resource, which works reliably.