✅ adjust tests
TL;DR: when we mocked `navigationStart` in `mockClock`, some tests were
passing only by chance.

In browsers that don't support `PerformanceNavigationTiming`, we rely on
`performance.timing` and a `setTimeout` (see [code][1]). Because
`navigationStart` was changed to an incoherent value, the timing entry
was judged ["incomplete"][2] and ignored.

Now that we don't mock `navigationStart` anymore, the entry is judged
"complete", and so it's not dismissed anymore, triggering an extra
update on browsers that don't support `PerformanceNavigationTiming`.

[1]: https://github.com/DataDog/browser-sdk/blob/54272ca00c7ddf5c016fb477365c3fd1cb4736de/packages/rum-core/src/domain/view/viewMetrics/trackNavigationTimings.ts#L68
[2]: https://github.com/DataDog/browser-sdk/blob/54272ca00c7ddf5c016fb477365c3fd1cb4736de/packages/rum-core/src/domain/view/viewMetrics/trackNavigationTimings.ts#L59