👷 allow job 'test-performance' to fail
This job randomly fail for various reason. While it is not required to
merge PRs at the GitHub level, it still blocks the gitlab pipeline, so
in practice it is required.

Because we only use this job to notify performance benchmark changes on
PRs, it is acceptable to bypass it sometimes. We'll need to figure out
how we can make this more stable in the future.