✅♻️ do not mock `navigationStart` in `mockClock` (#2979)
* ✅♻️ do not mock `navigationStart` in `mockClock`

Before this commit, `mockClock` was adjusting the `navigationStart`
value so calling `performance.now()` before any `clock.tick()` returned
0. This was useful for assertions on relative times, ex:

```js
clock = mockClock()
clock.tick(10)
expect(relativeNow()).toEqual(10)
```

Unfortunately, it lead to incoherences because some other timings were
not adjusted. For example `performance.timing.responseStart` was not
adjusted, so it was before `navigationStart`, which is unexpected.

To work around this, we locally tweak the mocked clock (ex: (here)[1]
and (here)[2]). Those tweaks are hard to understand, especially on
"integration tests" where those incoherences are deeply nested into
lower-level code.

So, this leaves us we two solutions to make things coherent again:

* Mock more things: replace values with hardcoded coherent values. This
  is tricky because more tests will need to mock more things. Also, we
  don't know where we'll need to stop mocking things, and at this rate
  is it still worth to execute the unit tests in real browsers?

* Mock less things: keep values provided by the browser

This commit chose the latter solution. We don't mock `navigationStart`
anymore, so `performance.now()` now returns an arbitrary value
influenced by the page load duration. To workaround this, the `Clock`
object now exposes helpers to compute times (relative time and
timestamp) relative to when `mockClock()` was invoked (a.k.a the start
of the test case). The above example should now look like this:

```js
clock = mockClock()
clock.tick(10)
expect(relativeNow()).toEqual(clock.relative(10))
```

[1]: https://github.com/DataDog/browser-sdk/blob/c6c0ad6125b825cb4755782a59e1593968707d55/packages/rum-core/src/domain/view/viewMetrics/trackNavigationTimings.spec.ts#L61
[2]: https://github.com/DataDog/browser-sdk/blob/c6c0ad6125b825cb4755782a59e1593968707d55/packages/rum-core/src/domain/view/trackViews.spec.ts#L436

* ✅ adjust tests

TL;DR: when we mocked `navigationStart` in `mockClock`, some tests were
passing only by chance.

In browsers that don't support `PerformanceNavigationTiming`, we rely on
`performance.timing` and a `setTimeout` (see [code][1]). Because
`navigationStart` was changed to an incoherent value, the timing entry
was judged ["incomplete"][2] and ignored.

Now that we don't mock `navigationStart` anymore, the entry is judged
"complete", and so it's not dismissed anymore, triggering an extra
update on browsers that don't support `PerformanceNavigationTiming`.

[1]: https://github.com/DataDog/browser-sdk/blob/54272ca00c7ddf5c016fb477365c3fd1cb4736de/packages/rum-core/src/domain/view/viewMetrics/trackNavigationTimings.ts#L68
[2]: https://github.com/DataDog/browser-sdk/blob/54272ca00c7ddf5c016fb477365c3fd1cb4736de/packages/rum-core/src/domain/view/viewMetrics/trackNavigationTimings.ts#L59