🐛 [RUM-5785] fix missing navigation timings on Safari (#2964)
* ✅♻️ move `mockDocumentReadyState` in a common module

And add the possibility to mock onload

* ✅ clone emitted view events

Sometimes when checking for View events emitted in the past, the
properties don't reflect the state at the time because they have been
mutated in place.

* ✅♻️ factorize the way we mock `window.performance.getEntries*`

* ♻️ make runOnReadyState stoppable

* ♻️ extract the code to polyfill "navigation" performance entry

We'll use this in `trackNavigationTimings` next.

* 🐛 [RUM-5785] fix track navigation timings on Safari

In most browsers, when observing `PerformanceNavigationTiming` via a
`PerformanceObserver` before the page is fully loaded, the same
"navigation" entry is notified twice:

* once with the "buffered" entry, even if it is incomplete

* once when all the timings are defined, after the `load` event is
  complete.

On Safari, the entry is only notified once. The SDK is only aware of the
incomplete entry (which is discarded), so all timings related to the
navigation entry are missing.

To work around this issue and keep things simple, this commit removes
the `PerformanceObserver` usage. Instead, we wait for the `load` event
to be complete, and then just pick up the buffered navigation entry.
This is similar to the strategy we use to generate the "document"
resource, which works reliably.

* ✅ adjust related integration tests

* 👌 review feedbacks

* ♻️  NavigationTiming inherits from ResourceTiming

The native PerformanceNavigationTiming is inheriting from
PerformanceResourceTiming, so let's do the same for our interfaces, so
we don't need to copy/paste/document the same properties.

* ✅ fix unit test in IE11