✨replace react-router wildcard routes with their actual path name (#3023)
* 👷 define TS paths for the react plugin

Before this commit, when running the dev-server, the react plugin was
taken from the built files (packages/rum-react/esm/....), so modifying
a source file did not re-build the sandbox.

Now, source files (from packages/rum-react/src/...) are used instead,
and modifying them trigger a sandbox rebuild.

* ✨ replace wildcard asterisks with the actual path

React Router routes defined with a '*' as a path are often used to
create custom routing logic. It doesn't bring much value within a View
name, as it could lead to very different pages. Replace them with the
actual path they are matching.