🐛 ignore layout shifts that happen before view start (#3058)
On "route change" views, the CLS computation took any buffered layout
shift entry into account, so in some cases it reflected the CLS of an
anterior view. This commit fixes the issue by ignoring layout shifts
that happened before the view start.

We still use the buffered entries so we can use layout shift entries
that happens before the SDK is initialized.