✅ set workerStart to 0 in mocked resource entries
`workerStart` defaults to 0 in resource entries, not 'undefined'. This
commit sets the correct default. It also removes the type cast, so we
have proper typechecking and we don't miss future properties in the
mock. To do this, I had to define all other properties so the entry is
actually complete.