🐛 skip worker timing when no worker is used (#3147)
* ✅ set workerStart to 0 in mocked resource entries

`workerStart` defaults to 0 in resource entries, not 'undefined'. This
commit sets the correct default. It also removes the type cast, so we
have proper typechecking and we don't miss future properties in the
mock. To do this, I had to define all other properties so the entry is
actually complete.

* ✅ simplify the test assertion

We don't need to re-check every single timings, just focus on the worker
one.

* 🐛 do not use undefined worker timings

`workerStart` defaults to 0 when there is no worker in use. Make sure it
is not 0 before collecting it.