✅ [RUM-6868] consider identifiers byte ordering little endian
This is not strictly needed from a functionality point of view, but
helps implementing tests that work with both the current
UintArray-based identifier and the future BigInt-based implementation.
The implementation is also simpler.

Rational:

In our tests, we are mocking `getRandomValues` to edit the
random bytes of the identifier, then we make sure the identifier gets
formated as a string correctly.

Currently, we set the last byte to 0xff, and since we consider this byte
to be the least significant byte (big endian), naturally the output is ff.

But in the next commit we'll introduce an alternative implementation of
identifiers based on BigInt generated using a BigUInt64Array. In this
setup, bigints are usually encoded as little endian. So if we edit the
last byte to 0xff, we actually edit the most significant byte, and the
output would be ff00000000000.

To keep tests simple, let's use the same endianness for both
implementation, so editing the same byte results in the same trace id.