⚗️ ✨ [RUM-6868] implement consistent probabilistic trace sampling  (#3186)
* 🚚 [RUM-6868] move crypto in a separate module

* 🚚 [RUM-6868] move identifier logic in a separate module

* 🚚 [RUM-6868] simplify identifier implementation

This commit reduces the scope of identifiers to make alternative
identifier implementations easier.

* ✨ [RUM-6868] make trace identifiers 64 bits

To comply with other tracers identifiers and be able to properly
implement probabilistic sampling, let's make trace identifiers 64 bits.

This PR differentiate Span and Trace identifiers to make sure we use
each at the right place.

* ✅ [RUM-6868] consider identifiers byte ordering little endian

This is not strictly needed from a functionality point of view, but
helps implementing tests that work with both the current
UintArray-based identifier and the future BigInt-based implementation.
The implementation is also simpler.

Rational:

In our tests, we are mocking `getRandomValues` to edit the
random bytes of the identifier, then we make sure the identifier gets
formated as a string correctly.

Currently, we set the last byte to 0xff, and since we consider this byte
to be the least significant byte (big endian), naturally the output is ff.

But in the next commit we'll introduce an alternative implementation of
identifiers based on BigInt generated using a BigUInt64Array. In this
setup, bigints are usually encoded as little endian. So if we edit the
last byte to 0xff, we actually edit the most significant byte, and the
output would be ff00000000000.

To keep tests simple, let's use the same endianness for both
implementation, so editing the same byte results in the same trace id.

* ✨ [RUM-6868] implement trace ids via bigints

* ✨ [RUM-6868] implement consistent trace sampling

* 🚩 [RUM-6868] put consistent sampling behind an experimental flag

* 👌 add comment about ie11 support

---------

Co-authored-by: Thomas Lebeau <thomas.lebeau@datadoghq.com>