🐛 fix document resource duration
Before this commit, we used the navigation entry duration as the
duration for the document resource. However, that duration includes the
time it takes to load the page until the `load` event (including the
load of assets declared in the HTML like JS, CSS, images...).

This commit fixes this by adjusting the duration to `responseEnd`, which
conforms with the duration of other resource entries.