🐛 fix document resource duration (#3307)
* ✅ improve performance entry fixture

* make sure all properties are defined (enforced by typechecking)
* always define `toJSON`
* adjust "Navigation" entry fixture

* ✅ consolidate tests a bit

* 🐛 fix navigation entry polyfill

In PerformanceNavigationTiming entries, the duration should reflect the
whole time it takes to load the page until the load event.

This is more for correctness than anything as we won't use navigation
entry duration anymore.

* 🐛 fix document resource duration

Before this commit, we used the navigation entry duration as the
duration for the document resource. However, that duration includes the
time it takes to load the page until the `load` event (including the
load of assets declared in the HTML like JS, CSS, images...).

This commit fixes this by adjusting the duration to `responseEnd`, which
conforms with the duration of other resource entries.