refactor: only depend on @openfeature/web-sdk
OpenFeature recommends relying on web-sdk package instead of core.

The downside is that web-sdk is not side-effect free, so it's better
if we don't depend on StandardResolutionReasons. The good news is that
TypeScript should be able to check literals against enum and using
'DEFAULT' literal is shorter and reduces bundle size a bit.