🔥 🐛 fix Next.js 12 + swcMinify compatibility
Next.js 12 introduced the `swcMinify` to opt-in using SWC to minify JS.
The SWC version used at the time was very early and suffered from major
issues. In particular, it was observed that the `computeSessionState`
function in `rumSessionManager.ts` was minified incorrectly in such a
way that it returned `undefined` every time.

This commit fixes this issue by adjusting the faulty function
implementation.

Instead of returning a "session state", it only returns the "session
tracking type" and consider that all sessions with tracking type '0' are
not tracked.

This reduces the bundle size a tiny bit, and prevents the concept of
"session state" to leak outside of `sessionStore`.