🔈 [RUM-10284] collect telemetry before starting the session manager (#3602)
* ✅ use fake telemetry in customerDataTelemetry

* ✅ use fake telemetry in rumInternalContext

* 🚚 add logs telemetry configuration in preStartLogs

This is to be uniform with RUM

* 🚚 move Logs telemetry transport logic to core

* 🔥 do not rely on session manager to flush telemetry

We don't really need to flush telemetry on session expiration. Let's
break this requirement, it will allow us to start and send telemetry
before starting the session manager.

* ♻️ allow setting telemetry context property by property

This is needed to progressively set context and don't crash if context
variables (ex: `session`) isn't defined yet when collecting telemetry.

* ♻️ allow passing a `createEncoder` to telemetry transport

* 🚚 start transport and drain in `startTelemetry`

* ✅ adjust e2e test as telemetry is sent separately

* 🚚 start logs telemetry before session manager

* 👌 fix typo

* ✅ add an e2e test

* 👌 add a unit test for progressive telemetry context